✨(backend) create backend docker image
The official moodlnet backend docker image is based on the the s6
process manager (via the s6-overlay project). It cannot be ran with a
non-privilegied user out of the box, due to the s6-overlay v1.22.1.0
that is used to build the image.

In addition to that, the image embeds a nginx server + the Erlang BEAM
application, supervised by s6. We want to have a container that run
_only_ the BEAM application.

And as a bonus, we prefer to use a debian based distrubution instead
of alpine.

The goal of this docker image is to address these concerns.