Deploy front application from circleci (#5)
* ✨(terraform) create configuration for state bucket

To be used in a team and never loose the terraform state, we create a
dedicated bucket and dynamodb instance to store the state and share it.
The dynamodb db is here to manage a lock on the state.

* ♻️(bin) allow to build frontend service from build script

build script was specific to build backend service. We modify it to also
build frontend service. A new argument is used to determine if it's
frontend or backend service to build, here is the new usage:

Usage: bin/build SERVICE [TAG]

OPTIONS:
  -h, --help   print this message

ARGUMENTS:
  SERVICE    the service to build. Must be backend or frontend
  TAG        image tag in the format name:tag

* ✨(terraform) create configuration to deploy frontend app

To deploy frontend app we choose to push files on a S3 bucket and serve
files with cloudfront using a custom domain. We need for that a use with
restricted access to s3 and cloudfront, a certificate for cloudfront
distribution and then a S3 bucket and a cloudfront distribution.

* 👷(circleci) configure a job a deploy frontend application

frontend application must be built in circleci and then deploy on S3